+++ designer/client/src/components/graph/node-modal/NodeTypeDetailsContent.tsx
@@ -1,10 +1,9 @@
 import { cloneDeep, isEqual, set } from "lodash";
 import type { SetStateAction } from "react";
-import React, { useCallback, useEffect, useMemo } from "react";
+import React, { useCallback, useEffect, useMemo, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 
-import { nodeDetailsClosed, nodeDetailsOpened, validateNodeData } from "../../../actions/nk";
-import { removeHistorySnapshot, takeHistorySnapshot } from "../../../reducers/graph/historySquash";
+import { validateNodeData } from "../../../actions/nk";
 import { getCreatorType } from "../../../reducers/selectors/getCreator";
 import { getProcessDefinitionData } from "../../../reducers/selectors/processDefinitionData";
 import type { Edge, NodeType, NodeValidationError } from "../../../types";
@@ -46,6 +45,19 @@ export type NodeTypeDetailsContentProps = {
     errors: NodeValidationError[];
 };
 
+export function useNodeAdjust() {
+    const getParameterDefinitions = useSelector(getDynamicParameterDefinitions);
+    const adjustNode = useCallback(
+        (node: NodeType) => {
+            const parameterDefinitions = getParameterDefinitions(node);
+            const { adjustedNode } = adjustParameters(node, parameterDefinitions);
+            return generateUUIDs(adjustedNode, ["fields", "parameters"]);
+        },
+        [getParameterDefinitions],
+    );
+    return adjustNode;
+}
+
 export function useNodeTypeDetailsContentLogic(props: Pick<NodeTypeDetailsContentProps, "onChange" | "node" | "edges" | "showValidation">) {
     const { onChange, node, edges, showValidation } = props;
     const dispatch = useDispatch();
@@ -60,6 +72,16 @@ export function useNodeTypeDetailsContentLogic(props: Pick<NodeTypeDetailsConten
 
     const variableTypes = useMemo(() => findAvailableVariables?.(node.id), [findAvailableVariables, node.id]);
 
+    const adjustNode = useNodeAdjust();
+    const [proxyNode, setProxyNode] = useState(() => adjustNode(node));
+
+    useEffect(() => {
+        setProxyNode((node) => {
+            const adjustedNode = adjustNode(node);
+            return isEqual(adjustedNode, node) ? node : adjustedNode;
+        });
+    }, [adjustNode]);
+
     const change = useCallback(
         (node: SetStateAction<NodeType>, edges: SetStateAction<Edge[]>) => {
             if (isEditMode) {
@@ -69,19 +91,20 @@ export function useNodeTypeDetailsContentLogic(props: Pick<NodeTypeDetailsConten
         [isEditMode, onChange],
     );
 
-    const setEditedNode = useCallback((n: SetStateAction<NodeType>) => change(n, edges), [edges, change]);
+    const setEditedNode = useCallback(
+        (n: SetStateAction<NodeType>) => {
+            setProxyNode((current) => {
+                const nextNode = typeof n === "function" ? n(current) : n;
+                change(nextNode, edges);
+                return nextNode;
+            });
+        },
+        [edges, change],
+    );
 
     const setEditedEdges = useCallback((e: SetStateAction<Edge[]>) => change(node, e), [node, change]);
 
     const parameterDefinitions = useMemo(() => getParameterDefinitions(node), [getParameterDefinitions, node]);
-
-    const adjustNode = useCallback(
-        (node: NodeType) => {
-            const { adjustedNode } = adjustParameters(node, parameterDefinitions);
-            return generateUUIDs(adjustedNode, ["fields", "parameters"]);
-        },
-        [parameterDefinitions],
-    );
 
     const renderFieldLabel = useCallback(
         (paramName: string): JSX.Element => {
@@ -136,10 +159,6 @@ export function useNodeTypeDetailsContentLogic(props: Pick<NodeTypeDetailsConten
         }
     }, [dispatch, edges, getBranchVariableTypes, node, processName, processProperties, showValidation, variableTypes]);
 
-    const adjustedNode = useMemo(() => {
-        return adjustNode(node);
-    }, [adjustNode, node]);
-
     return {
         ...props,
         isEditMode,
@@ -152,7 +171,7 @@ export function useNodeTypeDetailsContentLogic(props: Pick<NodeTypeDetailsConten
         removeElement,
         addElement,
         setProperty,
-        node: adjustedNode,
+        node: proxyNode,
     };
 }