+++ engine/flink/components/base-unbounded/src/main/scala/pl/touk/nussknacker/engine/flink/util/transformer/aggregate/aggregates.scala
@@ -223,7 +223,7 @@ object aggregates {
     override def result(finalAggregate: Aggregate): AnyRef = finalAggregate
 
     override def computeOutputType(input: typing.TypingResult): Validated[String, typing.TypingResult] = {
-      if (input.canBeConvertedTo(Typed[Boolean])) {
+      if (input.canBeLooselyAssignedTo(Typed[Boolean])) {
         Valid(Typed[Long])
       } else {
         Invalid(s"Invalid aggregate type: ${input.display}, should be: ${Typed[Boolean].display}")
@@ -483,7 +483,7 @@ object aggregates {
     ): Validated[String, TypedObjectTypingResult] = {
       input match {
         case TypedObjectTypingResult(inputFields, klass, _)
-            if inputFields.keySet == scalaFields.keySet && klass.canBeConvertedTo(
+            if inputFields.keySet == scalaFields.keySet && klass.canBeLooselyAssignedTo(
               Typed[java.util.Map[String, _]]
             ) =>
           val validationRes = scalaFields
@@ -569,7 +569,7 @@ object aggregates {
   trait MathAggregator { self: ReducingAggregator =>
 
     override def computeOutputType(input: typing.TypingResult): Validated[String, typing.TypingResult] = {
-      if (input.canBeConvertedTo(Typed[Number])) {
+      if (input.canBeLooselyAssignedTo(Typed[Number])) {
         // In some cases type can be promoted to other class e.g. Byte is promoted to Int for sum
         Valid(promotionStrategy.promoteSingle(input))
       } else {
@@ -597,7 +597,7 @@ object aggregates {
 
     override def computeOutputType(input: typing.TypingResult): Validated[String, typing.TypingResult] = {
 
-      if (!input.canBeConvertedTo(Typed[Number])) {
+      if (!input.canBeLooselyAssignedTo(Typed[Number])) {
         Invalid(s"Invalid aggregate type: ${input.display}, should be: ${Typed[Number].display}")
       } else {
         Valid(ForLargeFloatingNumbersOperation.promoteSingle(input))