+++ scenario-compiler/src/main/scala/pl/touk/nussknacker/engine/spel/Typer.scala
@@ -15,6 +15,7 @@ import pl.touk.nussknacker.engine.api.Context
 import pl.touk.nussknacker.engine.api.context.ValidationContext
 import pl.touk.nussknacker.engine.api.expression._
 import pl.touk.nussknacker.engine.api.generics.ExpressionParseError
+import pl.touk.nussknacker.engine.api.typed.ConversionStrategy.NoConversion
 import pl.touk.nussknacker.engine.api.typed.supertype.{CommonSupertypeFinder, NumberTypesPromotionStrategy}
 import pl.touk.nussknacker.engine.api.typed.typing._
 import pl.touk.nussknacker.engine.api.typed.typing.Typed.typedListWithElementValues
@@ -24,13 +25,7 @@ import pl.touk.nussknacker.engine.dict.SpelDictTyper
 import pl.touk.nussknacker.engine.expression.NullExpression
 import pl.touk.nussknacker.engine.spel.SpelExpressionParseError.{ArgumentTypeError, PartTypeError}
 import pl.touk.nussknacker.engine.spel.SpelExpressionParseError.IllegalOperationError._
-import pl.touk.nussknacker.engine.spel.SpelExpressionParseError.MissingObjectError.{
-  ConstructionOfUnknown,
-  NonReferenceError,
-  NoPropertyError,
-  NoPropertyTypeError,
-  UnresolvedReferenceError
-}
+import pl.touk.nussknacker.engine.spel.SpelExpressionParseError.MissingObjectError._
 import pl.touk.nussknacker.engine.spel.SpelExpressionParseError.OperatorError._
 import pl.touk.nussknacker.engine.spel.SpelExpressionParseError.SelectionProjectionError.{
   IllegalProjectionError,
@@ -148,8 +143,8 @@ private[spel] class Typer(
     def withChildrenOfType[Parts: universe.TypeTag](result: TypingResult) = {
       val w = valid(result)
       withTypedChildren {
-        case list if list.forall(_.canBeConvertedTo(Typed.fromDetailedType[Parts])) => w
-        case _                                                                      => w.tell(List(PartTypeError))
+        case list if list.forall(_.canBeLooselyAssignedTo(Typed.fromDetailedType[Parts])) => w
+        case _                                                                            => w.tell(List(PartTypeError))
       }
     }
 
@@ -200,7 +195,7 @@ private[spel] class Typer(
             case (ref: PropertyOrFieldReference) :: Nil => typeFieldNameReferenceOnRecord(ref.getName, record)
             case _                                      => typeFieldNameReferenceOnRecord(indexString, record)
           }
-        case indexKey :: Nil if indexKey.canBeConvertedTo(Typed[String]) =>
+        case indexKey :: Nil if indexKey.canBeLooselyAssignedTo(Typed[String]) =>
           if (dynamicPropertyAccessAllowed) valid(Unknown)
           else
             record.runtimeObjType.params match {
@@ -233,7 +228,7 @@ private[spel] class Typer(
             // It would be hard to change implementation of .asMap extension so we partially turn off this feature of indexer conversion
             // by allowing in typing only situations when map key type and indexer type are the same (though we have to allow
             // indexing with unknown type)
-            case indexKey :: Nil if indexKey.canBeConvertedWithoutConversionTo(keyParam) => valid(valueParam)
+            case indexKey :: Nil if indexKey.canBeAssignedTo(keyParam)(NoConversion) => valid(valueParam)
             case _ => invalid(IllegalIndexingOperation)
           }
         case d: TypedDict                    => dictTyper.typeDictValue(d, e).map(toNodeResult)
@@ -373,7 +368,8 @@ private[spel] class Typer(
 
       case e: OpMinus =>
         withTypedChildren {
-          case left :: right :: Nil if left.canBeConvertedTo(Typed[Number]) && right.canBeConvertedTo(Typed[Number]) =>
+          case left :: right :: Nil
+              if left.canBeLooselyAssignedTo(Typed[Number]) && right.canBeLooselyAssignedTo(Typed[Number]) =>
             val fallback = NumberTypesPromotionStrategy.ForMathOperation.promote(left, right)
             operationOnTypesValue[Number, Number, Number](left, right, fallback)((n1, n2) =>
               Valid(MathUtils.minus(n1, n2))
@@ -382,7 +378,7 @@ private[spel] class Typer(
             invalid(OperatorNonNumericError(e.getOperatorName, left))
           case left :: right :: Nil =>
             invalid(OperatorMismatchTypeError(e.getOperatorName, left, right))
-          case left :: Nil if left.canBeConvertedTo(Typed[Number]) =>
+          case left :: Nil if left.canBeLooselyAssignedTo(Typed[Number]) =>
             val resultType = left.withoutValue
             val result     = operationOnTypesValue[Number, Number](left)(MathUtils.negate).getOrElse(resultType)
             valid(result)
@@ -412,18 +408,20 @@ private[spel] class Typer(
         withTypedChildren {
           case left :: right :: Nil if left == Unknown || right == Unknown =>
             valid(Unknown)
-          case left :: right :: Nil if left.canBeConvertedTo(Typed[String]) || right.canBeConvertedTo(Typed[String]) =>
+          case left :: right :: Nil
+              if left.canBeLooselyAssignedTo(Typed[String]) || right.canBeLooselyAssignedTo(Typed[String]) =>
             operationOnTypesValue[Any, Any, String](left, right, Typed[String])((l, r) =>
               Valid(l.toString + r.toString)
             )
-          case left :: right :: Nil if left.canBeConvertedTo(Typed[Number]) && right.canBeConvertedTo(Typed[Number]) =>
+          case left :: right :: Nil
+              if left.canBeLooselyAssignedTo(Typed[Number]) && right.canBeLooselyAssignedTo(Typed[Number]) =>
             val fallback = NumberTypesPromotionStrategy.ForMathOperation.promote(left, right)
             operationOnTypesValue[Number, Number, Number](left, right, fallback)((n1, n2) =>
               Valid(MathUtils.plus(n1, n2))
             )
           case left :: right :: Nil =>
             invalid(OperatorMismatchTypeError(e.getOperatorName, left, right))
-          case left :: Nil if left.canBeConvertedTo(Typed[Number]) =>
+          case left :: Nil if left.canBeLooselyAssignedTo(Typed[Number]) =>
             valid(left)
           case left :: Nil =>
             invalid(OperatorNonNumericError(e.getOperatorName, left))
@@ -465,7 +463,7 @@ private[spel] class Typer(
           elementType <- extractIterativeType(iterateType)
           selectionType = resolveSelectionTypingResult(e, iterateType, elementType)
           result <- typeChildren(validationContext, node, current.pushOnStack(elementType)) {
-            case result :: Nil if result.canBeConvertedTo(Typed[Boolean]) =>
+            case result :: Nil if result.canBeLooselyAssignedTo(Typed[Boolean]) =>
               valid(selectionType)
             case other =>
               invalid(IllegalSelectionTypeError(other), selectionType)
@@ -476,7 +474,7 @@ private[spel] class Typer(
           case condition :: onTrue :: onFalse :: Nil =>
             for {
               _ <- Option(condition)
-                .filter(_.canBeConvertedTo(Typed[Boolean]))
+                .filter(_.canBeLooselyAssignedTo(Typed[Boolean]))
                 .map(valid)
                 .getOrElse(invalid(TernaryOperatorNotBooleanError(condition)))
             } yield CommonSupertypeFinder.Default.commonSupertype(onTrue, onFalse)
@@ -537,10 +535,10 @@ private[spel] class Typer(
       // as properly determining it would require evaluating the selection expression for each element (likely working on the AST)
       parentType match {
         case tc: SingleTypingResult
-            if tc.runtimeObjType.canBeConvertedTo(Typed[java.util.Collection[_]]) ||
+            if tc.runtimeObjType.canBeLooselyAssignedTo(Typed[java.util.Collection[_]]) ||
               tc.runtimeObjType.klass.isArray =>
           tc.withoutValue
-        case tc: SingleTypingResult if tc.runtimeObjType.canBeConvertedTo(Typed[java.util.Map[_, _]]) =>
+        case tc: SingleTypingResult if tc.runtimeObjType.canBeLooselyAssignedTo(Typed[java.util.Map[_, _]]) =>
           Typed.record(Map.empty)
         case _ =>
           parentType
@@ -592,7 +590,8 @@ private[spel] class Typer(
       op: Option[(Number, Number) => Validated[ExpressionParseError, Any]]
   )(implicit numberPromotionStrategy: NumberTypesPromotionStrategy): TypingR[CollectedTypingResult] = {
     typeChildren(validationContext, node, current) {
-      case left :: right :: Nil if left.canBeConvertedTo(Typed[Number]) && right.canBeConvertedTo(Typed[Number]) =>
+      case left :: right :: Nil
+          if left.canBeLooselyAssignedTo(Typed[Number]) && right.canBeLooselyAssignedTo(Typed[Number]) =>
         val fallback = numberPromotionStrategy.promote(left, right)
         op
           .map(operationOnTypesValue[Number, Number, Any](left, right, fallback)(_))
@@ -611,7 +610,7 @@ private[spel] class Typer(
       current: TypingContext
   )(op: Number => Any): TypingR[CollectedTypingResult] = {
     typeChildren(validationContext, node, current) {
-      case left :: Nil if left.canBeConvertedTo(Typed[Number]) =>
+      case left :: Nil if left.canBeLooselyAssignedTo(Typed[Number]) =>
         val result = operationOnTypesValue[Number, Any](left)(op).getOrElse(left.withoutValue)
         valid(result)
       case left :: Nil =>
@@ -707,10 +706,10 @@ private[spel] class Typer(
 
   private def extractIterativeType(parent: TypingResult): TypingR[TypingResult] = parent match {
     case tc: SingleTypingResult
-        if tc.runtimeObjType.canBeConvertedTo(Typed[java.util.Collection[_]]) ||
+        if tc.runtimeObjType.canBeLooselyAssignedTo(Typed[java.util.Collection[_]]) ||
           tc.runtimeObjType.klass.isArray =>
       valid(tc.runtimeObjType.params.headOption.getOrElse(Unknown))
-    case tc: SingleTypingResult if tc.runtimeObjType.canBeConvertedTo(Typed[java.util.Map[_, _]]) =>
+    case tc: SingleTypingResult if tc.runtimeObjType.canBeLooselyAssignedTo(Typed[java.util.Map[_, _]]) =>
       valid(
         Typed.record(
           Map(