+++ docs/Changelog.md
@@ -155,8 +155,12 @@
 * [#7553](https://github.com/TouK/nussknacker/pull/7553) Key variable created in window components is not transformed into a string.
   Key will now have type of value given in `groupBy` field. Previously we wrapped `groupBy` value as a list in String, e.g. for `id = 1` before: `groupBy: id -> key: "[1]"`, after: `groupBy: id -> key: 1`
 * [#7824](https://github.com/TouK/nussknacker/pull/7824) Add Json type support for schemaless topic data with Json content type.
-* [#7970](https://github.com/TouK/nussknacker/pull/7970) Added "activeScenariosLimit" setting defined per processing type. 
-
+* [#7959](https://github.com/TouK/nussknacker/pull/7959) Scenario testing API changes:
+    * Scenario testing API (on path prefix `/scenarioTesting`) is refactored - incompatible change, please check the Migration Guide for details
+    * test endpoints moved from Process Management API to Scenario Testing API
+    * introduced new representation of test results
+* [#7970](https://github.com/TouK/nussknacker/pull/7970) Added "activeScenariosLimit" setting defined per processing type
+* 
 ## 1.18
 
 #### Highlights